+++ .github/workflows/shortenertest.yml
@@ -56,7 +56,7 @@ jobs:
         run: |
           go install golang.org/x/tools/cmd/goimports@latest
 
-      - name: Build server binary
+      - name: Build shortener binary
         run: |
           cd cmd/shortener
           go build -buildvcs=false -o shortener
