+++ .gitignore
@@ -12,8 +12,8 @@
 *.out
 
 # Binaries
-cmd/server/main
-cmd/server/shortener
+cmd/shortener/main
+cmd/shortener/shortener
 
 # Dependency directories (remove the comment below to include it)
 vendor/
