--- cmd/server/README.md
@@ -1,3 +1,0 @@
-# cmd/shortener
-
-В данной директории будет содержаться код, который скомпилируется в бинарное приложение
