--- internal/app/README.md
@@ -1,3 +1,0 @@
-# internal/app
-
-В данной директории будет содержаться имплементация вашего сервиса
